requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.19.2
python-socks==2.0.1
requests==2.27.1
rsa==4.8
Telethon==1.24.0
urllib3==1.26.8
